navigation bar and navigation menu are now their own components, navigation is now a route instead of footer
