+++ src/routes/navigation/navigation.component.tsx
@@ -1,0 +1,19 @@
+import { Outlet } from "react-router-dom";
+
+import NavigationBar from "../../components/navigation-bar/navigation-bar.component";
+import NavigationMenu from "../../components/navigation-menu/navigation-menu.component";
+import Footer from "../../components/footer/footer.component";
+
+// Navigation bar appears on every page
+const Navigation = () => {
+    return (
+        <>
+            <NavigationBar />
+            <NavigationMenu />
+            <Outlet />
+            <Footer />
+        </>
+    );
+};
+
+export default Navigation;
